Stop dispatching SMIL events while seeking
Based on [1], 'beginEvent' and 'endEvent' should only be dispatched for
elements that are active at the seeked-to time and elements that were
active at the previous/old/seeked-from time but are inactive at the
seeked-to time (respectively). 'repeatEvent's should not be disaptched
while seeking.

To handle this, extend the TimingUpdate helper class with a map that
tracks what interval the active elements had before the seek operation,
and whether the timing update operation is a seek operation (called
"fast-forward" since we end up moving in the forward direction
regardless of the actual seek direction).
Apart from that, the dispatching of 'repeatEvent's are also inhibited,
as well as the notifications of repeats in general.

[1] https://www.w3.org/TR/SMIL3/smil-timing.html#q114

Bug: 1039886
Change-Id: I17d91b1355858ee0b916d4fa46b9572e685e40e5