[css-pseudo] Don't update non-normal markers if list-style-type changes
If a ::marker has the 'content' property set to something different than
'normal', then the contents of the marker should be determined according
to that value, not from 'list-style-type'.

Before this patch, the behavior was already correct when laying out the
marker for the first time. But if 'list-style-type' changed afterwards,
the marker was updated to reflect the new value, ignoring 'content'.

BUG=457718

TEST=external/wpt/css/css-pseudo/marker-content-021.html

The test fails in legacy since 'content' is not supported yet.

Change-Id: Id845774739927c9c574a66e9d1cc94e6ea59d5e5